@@ -8,22 +8,22 @@ import java.util.UUID;
 
 public class PitPlayerManager {
 
-    private static final Map<UUID, PitPlayer> pitPlayerHashMap = new HashMap<>();
-    public static void registerUser(Player player) {
-        pitPlayerHashMap.put(player.getUniqueId(), new PitPlayer(player));
+    private static final Map<UUID, PitPlayer> pitPlayerMap = new HashMap<>();
+    public static void registerUser(PitPlayer player) {
+        pitPlayerMap.put(player.getUniqueId(), player);
     }
 
-    public static void unregisterUser(Player player) {
+    public static void unregisterUser(PitPlayer player) {
         final UUID uuid = player.getUniqueId();
-        pitPlayerHashMap.remove(uuid);
+        pitPlayerMap.remove(uuid);
     }
 
     public static PitPlayer getPitPlayer(Player player) {
-        return pitPlayerHashMap.get(player.getUniqueId());
+        return pitPlayerMap.get(player.getUniqueId());
     }
 
     public static boolean isContain(Player player) {
-        return pitPlayerHashMap.containsKey(player.getUniqueId());
+        return pitPlayerMap.containsKey(player.getUniqueId());
     }
 
 }
